requirements.txt
requests
requests-aws4auth
python-dotenv
paho-mqtt
opencv-python
numpy
ultralytics
